+++ inc/initfs.h
@@ -6,6 +6,6 @@
 /* Scan multiboot2 tags for a module with name `module_name`.
    If found, unpack cpio newc archive from the module into the VFS.
    Returns 0 on success, negative on error, 1 if module not found. */
-int initfs_process_multiboot_module(uint32_t multiboot_magic, uint32_t multiboot_info, const char *module_name);
+int initfs_process_multiboot_module(uint32_t multiboot_magic, uint64_t multiboot_info, const char *module_name);
